s6-dns: add new recipe
Introduce s6-dns, a suite of DNS client programs and libraries for Unix
systems, as an alternative to the BIND, djbdns or other DNS clients.

It's required for s6-networking which is highly recommended for s6 socket
activation when doing supervision with s6.

See https://skarnet.org/software/s6-dns/ for details.

Signed-off-by: Jens Rehsack <sno@netbsd.org>